refactor(in_list): Extract ArrayStaticFilter to in_list/array_filter.rs
Extract the generic hash-based filter.

- Create in_list/array_filter.rs with ArrayStaticFilter
- Update in_list.rs to use the filter from the submodule
- Remove duplicate ArrayStaticFilter definition from main file

ArrayStaticFilter is the fallback filter for complex types that don't
have specialized implementations (structs, nested types, etc.).
